6 tests left

+++ lib/libft.h
@@ -6,7 +6,7 @@
 /*   By: cghanime <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 20:00:22 by cghanime          #+#    #+#             */
-/*   Updated: 2019/08/28 06:06:00 by cghanime         ###   ########.fr       */
+/*   Updated: 2019/08/28 17:31:38 by cghanime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,7 +82,7 @@ char				*ft_strjoinfrchoz(char *s1, char *s2, int choice);
 char				*ft_strtrim(char *s);
 char				**ft_strsplit(char *s, char c);
 char				*ft_itoa(int n);
-char				*ft_ltoa(long long n);
+char				*ft_ltoa(intmax_t n);
 char				*ft_lltoa(long long n);
 void				ft_putchar(char c);
 void				ft_putstr(char *s);
